requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
facebook-sdk==2.0.0
Flask==0.12.2
Flask-Login==0.4.0
Flask-SocketIO==2.9.2
Flask-SQLAlchemy==2.3.0
Flask-WTF==0.14.2
gevent==1.2.2
gevent-websocket==0.10.1
google-api-python-client==1.6.3
greenlet==0.4.12
httplib2==0.10.3
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
oauth2==1.9.0.post1
oauth2client==4.1.2
pyasn1==0.3.3
pyasn1-modules==0.1.1
pyparsing==2.2.0
pyScss==1.3.5
python-engineio==1.7.0
python-socketio==1.8.0
redis==2.10.6
requests==2.18.4
rsa==3.4.2
six==1.10.0
SQLAlchemy==1.1.14
udatetime==0.0.13
uritemplate==3.0.0
urllib3==1.22
uWSGI==2.0.15
webassets==0.12.1
Werkzeug==0.12.2
WTForms==2.1
